Transitioned fare amount from `f64` to `rust_decimal::Decimal`.

+++ src/directions/response/transit_fare.rs
@@ -1,11 +1,12 @@
 use crate::directions::response::transit_currency::TransitCurrency;
+use rust_decimal::Decimal;
 use serde::{Serialize, Deserialize};
 
 /// If present, contains the total fare (that is, the total ticket costs) on
 /// this route. This property is only returned for transit requests and only for
 /// routes where fare information is available for all transit legs.
 
-#[derive(Clone, Debug, PartialEq, PartialOrd, Serialize, Deserialize)]
+#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize, Deserialize)]
 pub struct TransitFare {
 
     /// An [ISO 4217 currency code](https://en.wikipedia.org/wiki/ISO_4217)
@@ -16,6 +17,6 @@ pub struct TransitFare {
     pub text: String,
 
     /// The total fare amount, in the currency specified above.
-    pub value: f64,
+    pub value: Decimal,
 
 } // struct
